+++ example/async_/settings.py
@@ -3,3 +3,12 @@
 from ..settings import *  # noqa: F403
 
 ROOT_URLCONF = "example.async_.urls"
+
+CHANNEL_LAYERS = {
+    "default": {
+        "BACKEND": "channels_redis.core.RedisChannelLayer",
+        "CONFIG": {
+            "hosts": [("127.0.0.1", 6379)],
+        },
+    }
+}